matched func_8005D180_5DD80

+++ src/5DBC0.h
@@ -9,7 +9,7 @@ s32 getTrackHeightInSectorWithOffset(void *gameData, u16 sectorIndex, void *pos,
 void func_8005CFFC_5DBFC(void *arg0, u16 arg1, void *arg2, void *arg3, void *arg4);
 s32 getSectorTrackHeight(TrackGeometryFaceData *geom, u16 groupIdx, Vec3i *pos, s32 yOffset);
 void clampPlayerVelocityToMaxSpeed(Player *player);
-s32 func_8005D180_5DD80(void *arg0, s32 arg1);
+s32 func_8005D180_5DD80(void *arg0, s16 arg1);
 s32 func_8005D308_5DF08(void *arg0, s32 arg1);
 void setPlayerLeanAnimation(Player *player, s32 animIndex, s32 progress);
 void loadCharacterBodyParts(Player *player);
